All 5-card hands in an ascending order by hand value
assert. 2598960 5 -: $RH5c
   datatype RH5c    NB. Numbers 0-51 encoded as characters for speed of processing.
literal

For example, here are the 3 lowest and the 5 highest hands:
   showCards (3&{.,_5&{.)a. i. RH5c
+--+--+--+---+---+
|7♦|5♣|4♣|3♣ |2♣ |
+--+--+--+---+---+
|7♥|5♣|4♣|3♣ |2♣ |
+--+--+--+---+---+
|7♠|5♣|4♣|3♣ |2♣ |
+--+--+--+---+---+
|K♠|Q♠|J♠|10♠|9♠ |
+--+--+--+---+---+
|A♣|K♣|Q♣|J♣ |10♣|
+--+--+--+---+---+
|A♦|K♦|Q♦|J♦ |10♦|
+--+--+--+---+---+
|A♥|K♥|Q♥|J♥ |10♥|
+--+--+--+---+---+
|A♠|K♠|Q♠|J♠ |10♠|
+--+--+--+---+---+

"75..." is lowest, not "64..." as we would do in lowball, because "6" followed by four lower cards is either a straight or includes an ace as in lowball but ace is a high card in high poker.

Card values in each row are in ascending numerical value, as enforced by "strictord", e.g. 
   a. i. 3{.RH5c
0 1 2 3 18
0 1 2 3 31
0 1 2 3 44
   strictord 3 2 0 1 18
0 1 2 3 18